#6.1.1 The `InputKeyBoard.DoubleClick` property had an issue with remaining true while the mouse button was held down because the property was not reset correctly.
The `Gizmos` class no longer inherits from the `Node` class and has become static classes.
This was done so that these classes do not inherit methods or fields derived from the Node class.

The `Node_GD_CB_Extension.SetParent(this Node?, Node?)` extension method was added to the `Node` class.